chore: split store to separate units
