+++ lerobot-web-fe/src/components/CameraStream.tsx
@@ -1,16 +1,17 @@
 import { ExclamationTriangleIcon } from '@radix-ui/react-icons';
-import { useStatusStore } from '../stores/status.store';
+import { useConfigStore } from '../stores/config.store';
+import { useCameraStore } from '../stores/camera.store';
 import { CameraBox } from './CameraBox';
 
 import styles from './CameraStream.module.css';
 
 export function CameraStream() {
-  const apiUrl = useStatusStore((state) => state.apiUrl);
-  const cameraIds = useStatusStore((state) => state.cameraList?.cameras ?? []);
+  const apiUrl = useConfigStore((state) => state.apiUrl);
+  const cameraIds = useCameraStore((state) => state.cameraList?.cameras);
 
   return (
     <div className={styles.cameraStream}>
-      {cameraIds.length > 0 && apiUrl ? (
+      {cameraIds && cameraIds.length > 0 && apiUrl ? (
         cameraIds.map((cameraId) => (
           <div key={cameraId} className={styles.cameraWrapper}>
             <CameraBox apiUrl={apiUrl} id={cameraId} />
