+++ lerobot-web-fe/src/models/camera.model.ts
@@ -1,0 +1,3 @@
+export interface CameraListResponse {
+  cameras: number[];
+}
