+++ lerobot-web-fe/src/models/robot.model.ts
@@ -1,5 +1,3 @@
-import type { StatusResponse } from '../models/status.model';
-
 export const DEFAULT_ROBOT_COUNT = 2;
 export const EMPTY_ROBOT_INDEX = '-1';
 
@@ -51,37 +49,3 @@ export interface RobotStatus {
   name: string;
   robot_type: string;
 }
-
-export const mockStatusResponse: StatusResponse = {
-  status: 'ok',
-  name: 'jetson',
-  robots: ['so-100', 'so-100', 'agilex-piper'],
-  robot_status: [
-    {
-      name: 'so-100',
-      robot_type: 'manipulator',
-      device_name: '58FA101935',
-    },
-    {
-      name: 'so-100',
-      robot_type: 'manipulator',
-      device_name: '5A4B049137',
-    },
-    {
-      name: 'agilex-piper',
-      robot_type: 'manipulator',
-      device_name: null,
-    },
-  ],
-  cameras: {
-    cameras_status: [],
-    is_stereo_camera_available: false,
-    realsense_available: false,
-    video_cameras_ids: [],
-  },
-  version_id: '0.3.49',
-  is_recording: false,
-  ai_running_status: 'stopped',
-  server_ip: '192.168.100.102',
-  leader_follower_status: false,
-};
